tests: reorganise automated tests executed in prs for better visibility
