+++ test/e2e/tests/crtitical_tests_prs/__init__.py
@@ -1,0 +1,3 @@
+import pytest
+
+marks = pytest.mark.critical
