+++ test/e2e/tests/crtitical_tests_prs/test_community_category_add_edit_delete.py
@@ -1,0 +1,79 @@
+import allure
+import pytest
+from allure_commons._allure import step
+
+import constants
+from gui.main_window import MainWindow
+from . import marks
+
+pytestmark = marks
+
+
+@allure.testcase('https://ethstatus.testrail.net/index.php?/cases/view/703226', 'Add category')
+@allure.testcase('https://ethstatus.testrail.net/index.php?/cases/view/703227', 'Remove category')
+@allure.testcase('https://ethstatus.testrail.net/index.php?/cases/view/703233', 'Edit category title')
+@allure.testcase('https://ethstatus.testrail.net/index.php?/cases/view/703234', 'Edit category - add channel')
+@allure.testcase('https://ethstatus.testrail.net/index.php?/cases/view/703235', 'Edit category - remove channel')
+@pytest.mark.case(703226, 703233, 703234, 703235, 703227)
+@pytest.mark.parametrize(
+    'category_name, general_checkbox, channel_name, channel_description, channel_emoji, second_channel_name, '
+    'second_channel_description, second_channel_emoji',
+    [pytest.param('Category in general', True, 'Channel', 'Description', 'sunglasses', 'Second-channel',
+                  'Description', 'sunglasses')])
+@pytest.mark.critical
+def test_create_edit_remove_community_category(main_screen: MainWindow, category_name, general_checkbox, channel_name,
+                                 channel_description, channel_emoji, second_channel_name, second_channel_description,
+                                 second_channel_emoji):
+    with step('Enable creation of community option'):
+        settings = main_screen.left_panel.open_settings()
+        settings.left_panel.open_advanced_settings().enable_creation_of_communities()
+    with step('Create community and select it'):
+        community_params = constants.community_params
+        main_screen.create_community(community_params['name'], community_params['description'],
+                                     community_params['intro'], community_params['outro'],
+                                     community_params['logo']['fp'], community_params['banner']['fp'])
+        community_screen = main_screen.left_panel.select_community(community_params['name'])
+
+    with step('Create community category and verify that it displays correctly'):
+        community_screen.create_category(category_name, general_checkbox)
+
+    with step('Verify category'):
+        community_screen.verify_category(category_name)
+
+    with step('Create community channel inside category'):
+        community_screen.left_panel.open_new_channel_popup_in_category().create(channel_name, channel_description,
+                                                                                channel_emoji).save()
+
+    with step('Create community channel outside of category'):
+        community_screen.create_channel(second_channel_name, second_channel_description, second_channel_emoji)
+
+    with step('Verify that selected channel is listed outside of category'):
+        assert community_screen.left_panel.get_channel_or_category_index(second_channel_name) == 0
+
+    with step('Open edit category popup'):
+        category_popup = community_screen.edit_category()
+        category_popup.enter_category_title("New category").click_checkbox_by_index(0)
+        category_popup.save()
+
+    with step('Verify that selected channel is now listed inside category'):
+        assert community_screen.left_panel.get_channel_or_category_index(second_channel_name) == 3
+
+    with step('Open edit category popup'):
+        category_popup = community_screen.edit_category()
+        category_popup.click_checkbox_by_index(2)
+        category_popup.save()
+
+    with step('Verify that selected channel is now listed outside of category'):
+        assert community_screen.left_panel.get_channel_or_category_index(second_channel_name) == 0
+
+    with step('Delete category'):
+        community_screen.delete_category()
+
+    with step('Verify category is not in the list'):
+        assert category_name not in community_screen.left_panel.categories_items
+
+    with step('Verify created channel and general channel are still in the list'):
+        new_channel = community_screen.left_panel.get_channel_parameters(channel_name)
+        general_channel = community_screen.left_panel.get_channel_parameters('general')
+        assert new_channel in community_screen.left_panel.channels
+        assert general_channel in community_screen.left_panel.channels
