+++ test/e2e/tests/crtitical_tests_prs/test_onboarding_import_seed.py
@@ -4,7 +4,7 @@ from allure_commons._allure import step
 
 from constants.onboarding import KeysExistText
 from driver.aut import AUT
-from . import marks
+from tests.onboarding import marks
 
 import configs.system
 import constants
@@ -31,7 +31,6 @@ def keys_screen(main_window) -> KeysView:
 @pytest.mark.case(703040, 736372)
 @pytest.mark.parametrize('user_account', [constants.user.user_account_one])
 @pytest.mark.parametrize('autocomplete, default_name', [
-    pytest.param(False, 'Account 1'),
     pytest.param(True, 'Account 1', marks=pytest.mark.critical)
 ])
 def test_import_seed_phrase(keys_screen, main_window, aut: AUT, user_account, default_name: str, autocomplete: bool):
