+++ test/e2e/tests/crtitical_tests_prs/test_plus_button_add_account_after_restart.py
