+++ test/e2e/tests/crtitical_tests_prs/test_wallet_settings_acct_interactions_delete_account.py
@@ -4,7 +4,7 @@ import string
 import allure
 import pytest
 from allure_commons._allure import step
-from . import marks
+from tests.settings.settings_wallet import marks
 
 import constants
 import driver
