--- test/e2e/tests/user_guide_tests/__init__.py
@@ -1,3 +1,0 @@
-import pytest
-
-marks = pytest.mark.user_guide
